feat: updates upload process to use POST method
Switches the upload method to POST and includes organization and tenant IDs, as well as an authorization token, in the request.  This enables secure uploads with proper context.

Modifies the UploaderService to construct a config object with the URL and authentication details and passes it to the upload worker. The worker then uses this config to make the POST request with the necessary headers and form data.  If no upload URL is found from the upload service, fetches signed URL as before.