fix(electron-utils): get correct browser window for progress bar
The previous method of obtaining the browser window using
`BrowserWindow.getAllWindows()[0]` was unreliable. It could potentially
select the wrong window if multiple were open or fail if accessed
before a window was fully initialized.
This change introduces a getter that uses `BrowserWindow.fromId` with
the `event.frameId` to reliably retrieve the specific window associated
with the IPC event triggering the worker handler. Optional chaining is
added to `setProgressBar` calls for added safety.