feat(upload): honor auto-sync setting for automatic uploads
The upload process will now check the `autoSync` setting before
automatically starting the next item in the queue.
A new selector `selectSettingUploadAutoSync` has been added to retrieve
this setting from `web-setting-data-access`. The `startNextUpload$`
effect in upload data-access uses this selector, along with existing
checks for `canUploadMore` and items in the queue, to determine if the
next upload should proceed.
As part of this change, the `No Operation` action has been removed from
upload actions, as the effect now returns `EMPTY` if conditions for
auto-upload are not met.